Moved the logic to display the handles to an external file.
As inetended in previous commit.
